+++ toDoList/src/components/Header/index.js
@@ -6,7 +6,7 @@ import styles from './styles';
 const Header = () => {
   return (
     <View style={styles.box}>
-      <Text style={styles.welcome}>Todo List</Text>
+      <Text style={styles.welcome}>Piaccere</Text>
     </View>
   );
 };
