+++ toDoList/src/components/Navigation/tabNavigation/index.js
@@ -1,0 +1,11 @@
+// import React from 'react';
+import { createTabNavigator, createMaterialTopTabNavigator, createAppContainer } from 'react-navigation';
+import ToDo from '../../../screens/ToDo';
+import Books from '../../../screens/ZBooks';
+
+const TabNavigator = createMaterialTopTabNavigator({
+  Todo: { screen: ToDo },
+  Books: { screen: Books }
+});
+
+export default createAppContainer(TabNavigator);
