+++ toDoList/src/screens/ToDo/components/Row/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { gray_one, gray_two } from '../../constants/colors';
+import { gray_one, gray_two } from '../../../../constants/colors';
 
 export default StyleSheet.create({
   container: {
