+++ toDoList/src/screens/ToDo/components/ToDoInput/styles.js
@@ -1,6 +1,6 @@
 import { Platform, StyleSheet } from 'react-native';
 
-import { gray_three, gray_four } from '../../constants/colors';
+import { gray_three, gray_four } from '../../../../constants/colors';
 
 export default StyleSheet.create({
   header: {
