+++ toDoList/src/screens/ZBooks/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+
+import styles from './style';
+
+const Books = () => {
+ return (
+  <View style={styles.container}>
+   <Text style={styles.text}>Books!</Text>
+  </View>
+ )};
+
+export default Books;
