+++ toDoList/src/screens/ZBooks/style.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+// import { white, gray_three } from '../../constants/colors';
+
+export default StyleSheet.create({
+ container: {
+  flex: 1,
+  justifyContent: 'center',
+  alignItems: 'center'
+},
+ text: {
+  textAlign: 'center',
+  fontSize:30,
+},
+});
+
