+++ toDoList/styles.js
@@ -1,29 +1,10 @@
-import { Platform, StyleSheet } from 'react-native';
+import { StyleSheet } from 'react-native';
 
-import { white, gray_three } from './src/constants/colors';
+import { white } from './src/constants/colors';
 
 export default StyleSheet.create({
- container: {
-  flex: 1,
-  backgroundColor: white,
-},
-list: {
-  backgroundColor: white,
-  ...Platform.select({
-   ios:{
-    minHeight: 570,
-   },
-   android: {
-    minHeight: 460,
-   },
-  })
-},
-content: {
-  flex: 1,
-},
-separator: {
-  borderWidth: 1,
-  borderColor: gray_three,
-},
+  container: {
+    flex: 1,
+    backgroundColor: white
+  },
 });
-
